requirements.txt
Django==3.2.20
flake8==5.0.4
django-environ==0.10.0
Pillow==9.5.0
django_extensions==3.2.3
django-lifecycle==1.0.0
celery[redis]==5.2.7
django-celery-results==2.5.1
django-celery-beat==2.5.0
requests==2.31.0
pymemcache==4.0.0
phonenumbers==8.13.20
django-phonenumber-field==7.1.0
django-silk==5.0.3
pytest==8.3.3
pytest-django==4.9.0
faker==18.13.0
pprintpp==0.4.0
pytest-mock==3.11.1
coverage==7.2.7
pytest-factoryboy==2.6.0
beautifulsoup4==4.12.2
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
markdown==3.4.4
django-filter==23.5
psycopg2==2.9.0
gunicorn==23.0.0